Widen the weight products so the clamps above them can fire
Three places accumulate carried weight from count times weight per item. Two of
them -- TradeList.validate and Inventory.refreshWeight -- take the total in long
and finish with Math.min(weight, Integer.MAX_VALUE). That clamp is what gives
them away: it can only ever do anything if the product reaching the long is
itself computed in long, and it was computed in int, so it wrapped before the
clamp saw it and Math.min returned the wrapped value.

The third, TradeList.privateStoreBuy, took the total in int with no clamp at all,
while its sibling fifty lines above takes it in long.

Measured: adena weighs nothing, and the heaviest stackable item in the shipped
data weighs 1000, so the int product turns over at about 2.1 million units of it.
Past that line the reported weight goes negative and the carry limit stops
meaning anything for that character, which is exactly the state the clamp was
written to prevent.

This is the same shape already fixed in MultiSellChoose and RequestBuyItem, found
by sweeping the core and the datapack for a product handed to a long or a double
with no cast in front of it.

+++ java/org/l2jmobius/gameserver/model/itemcontainer/Inventory.java
@@ -1225,7 +1225,10 @@ public abstract class Inventory extends ItemContainer {
 		long weight = 0;
 		for (Item item : _items) {
 			if ((item != null) && (item.getTemplate() != null)) {
-				weight += item.getTemplate().getWeight() * item.getCount();
+				// Math.min below only means something if the product is widened first: taken
+				// in int it wraps before it reaches the long, and a wrapped total reports the
+				// character as carrying less than nothing.
+				weight += (long) item.getTemplate().getWeight() * item.getCount();
 			}
 		}
